@@ -30,13 +30,11 @@ public class AuthRequest extends Thread {
 
     @Override
     public void run() {
-        new CSRFGet(token -> {
-            Global.client.newCall(new Request.Builder().url(url)
-                .addHeader("Referer", referer)
-                .addHeader("X-CSRFToken", token)
-                .addHeader("X-Requested-With", "XMLHttpRequest")
-                .method(method, body)
-                .build()).enqueue(callback);
-        }, referer).start();
+        new CSRFGet(token -> Global.client.newCall(new Request.Builder().url(url)
+            .addHeader("Referer", referer)
+            .addHeader("X-CSRFToken", token)
+            .addHeader("X-Requested-With", "XMLHttpRequest")
+            .method(method, body)
+            .build()).enqueue(callback), referer).start();
     }
 }
