@@ -8,6 +8,7 @@ import android.graphics.Bitmap;
 import android.graphics.drawable.BitmapDrawable;
 import android.graphics.drawable.Drawable;
 import android.net.Uri;
+import android.os.Build;
 import android.view.Menu;
 import android.view.MenuItem;
 
@@ -33,17 +34,13 @@ import java.util.Random;
 
 public class Utility {
     public static final Random RANDOM = new Random(System.nanoTime());
-    public static final String ORIGINAL_URL = "nhentai.net";
-
     @NonNull
-    @Contract(pure = true)
     public static String getBaseUrl() {
         return "https://" + Utility.getHost() + "/";
     }
-
-
+    @NonNull
     public static String getHost() {
-        return Global.getMirror();
+        return "nhentai.net";
     }
 
     private static void parseEscapedCharacter(@NonNull Reader reader, Writer writer) throws IOException {
